Refactor Resource classes into a package
