+++ vendorflow-jsonapi-java-annotation-processor/src/main/resources/co/vendorflow/oss/jsonapi/processor/templates/JsonApiTypeRegistration.ftl
@@ -1,8 +1,13 @@
 package ${tr.packageName};
 
-@javax.annotation.Generated("co.vendorflow.oss.jsonapi.processor.TypeRegistrationProcessor")
-public final class ${tr.simpleName} implements co.vendorflow.oss.jsonapi.jackson.JsonApiTypeRegistration {
+import javax.annotation.Generated;
+
+import co.vendorflow.oss.jsonapi.jackson.JsonApiTypeRegistration;
+import co.vendorflow.oss.jsonapi.model.resource.JsonApiResource;
+
+@Generated("co.vendorflow.oss.jsonapi.processor.TypeRegistrationProcessor")
+public final class ${tr.simpleName} implements JsonApiTypeRegistration {
   @Override public String namespace() { return ""; }
   @Override public String typeName() { return "${tr.jsonApiType}"; }
-  @Override public Class<? extends co.vendorflow.oss.jsonapi.model.JsonApiResource<?, ?>> typeClass() { return ${tr.resourceFqcn}.class ; }
+  @Override public Class<? extends JsonApiResource<?, ?>> typeClass() { return ${tr.resourceFqcn}.class ; }
 }
