+++ vendorflow-jsonapi-java-annotation-processor/src/test/java/co/vendorflow/oss/jsonapi/processor/AlphaAttributes.java
@@ -1,6 +1,6 @@
 package co.vendorflow.oss.jsonapi.processor;
 
-import co.vendorflow.oss.jsonapi.model.JsonApiAttributes;
+import co.vendorflow.oss.jsonapi.model.resource.JsonApiAttributes;
 import lombok.Data;
 
 @JsonApiAttributes(type = "alphas")
