+++ vendorflow-jsonapi-java-annotation-processor/src/test/java/co/vendorflow/oss/jsonapi/processor/BravoAttributes.java
@@ -3,7 +3,7 @@ package co.vendorflow.oss.jsonapi.processor;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Size;
 
-import co.vendorflow.oss.jsonapi.model.JsonApiAttributes;
+import co.vendorflow.oss.jsonapi.model.resource.JsonApiAttributes;
 import lombok.Data;
 
 @JsonApiAttributes(type = "bravos", resourceSuffix = "Dto", meta = BravoAttributes.Meta.class)
