+++ vendorflow-jsonapi-java-annotation-processor/src/test/java/co/vendorflow/oss/jsonapi/processor/CharlieAttributes.java
@@ -2,7 +2,7 @@ package co.vendorflow.oss.jsonapi.processor;
 
 import java.util.List;
 
-import co.vendorflow.oss.jsonapi.model.JsonApiAttributes;
+import co.vendorflow.oss.jsonapi.model.resource.JsonApiAttributes;
 import lombok.Data;
 
 @JsonApiAttributes(type = "charlies", resourceTypeName = "ChuckRez", nullable = true)
