+++ vendorflow-jsonapi-java-annotation-processor/src/test/java/co/vendorflow/oss/jsonapi/processor/MikeResource.java
@@ -1,7 +1,7 @@
 package co.vendorflow.oss.jsonapi.processor;
 
-import co.vendorflow.oss.jsonapi.model.JsonApiType;
-import co.vendorflow.oss.jsonapi.model.MapAttributesJsonApiResource;
+import co.vendorflow.oss.jsonapi.model.resource.JsonApiType;
+import co.vendorflow.oss.jsonapi.model.resource.MapAttributesJsonApiResource;
 
 @JsonApiType("mikes")
 public class MikeResource extends MapAttributesJsonApiResource {
